[#24] refactor: 너무 큰 DetailView를 분리
