+++ SideDish/SideDish/View/DetailViewController.swift
@@ -2,11 +2,34 @@ import UIKit
 
 class DetailViewController: UIViewController {
 
-    private let detailView: DetailView
+    private let detailDishSectionView = DetailDishSectionView()
+    private let deliverySectionView = DeliverySectionView()
+    private let amountSectionView = AmountSectionView()
     private var viewModel: DishDetailViewModel?
 
+    private let mainScrollView: UIScrollView = {
+        var scrollView = UIScrollView()
+        scrollView.translatesAutoresizingMaskIntoConstraints = false
+        scrollView.showsVerticalScrollIndicator = false
+        return scrollView
+    }()
+
+    private let contentView: UIView = {
+        var view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+
+    private let detailImageStackView: UIStackView = {
+        var stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.distribution = .fill
+        stackView.alignment = .fill
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        return stackView
+    }()
+
     init(viewModel: DishDetailViewModel) {
-        detailView = DetailView()
         self.viewModel = viewModel
         super.init(nibName: nil, bundle: nil)
 
@@ -14,7 +37,6 @@ class DetailViewController: UIViewController {
     }
 
     required init?(coder: NSCoder) {
-        detailView = DetailView()
         self.viewModel = nil
         super.init(coder: coder)
     }
@@ -23,13 +45,18 @@ class DetailViewController: UIViewController {
         super.viewWillAppear(animated)
 
         viewModel?.fetchDetailData()
-        view = detailView
     }
 
     override func viewDidLoad() {
         super.viewDidLoad()
         setNavigationItems()
-        detailView.setLayout()
+        view.backgroundColor = .white
+
+        setMainScrollViewLayout()
+        setDetailDishSectionViewLayout()
+        setDeliverySectionViewLayout()
+        setAmountSectionViewLayout()
+        setDetailImageStackViewLayout()
     }
 
     private func setNavigationItems() {
@@ -46,48 +73,19 @@ class DetailViewController: UIViewController {
 
     private func setOnUpdate() {
         viewModel?.onUpdate = {
-            DispatchQueue.main.async {
-                self.navigationItem.title = self.viewModel?.title
-                self.detailView.titleLabel.text = self.viewModel?.title
-                self.detailView.descriptionLabel.text = self.viewModel?.description
-                self.detailView.finalPriceLabel.text = self.viewModel?.finalPrice
-                self.detailView.pointLabel.text = self.viewModel?.point
-                self.detailView.shippingInfoLabel.text = self.viewModel?.deliveryInfo
-                self.detailView.shippingFeeLabel.text = self.viewModel?.deliveryFee
-
-                if let normalPrice = self.viewModel?.normalPrice {
-                    let attributedStyle = NSUnderlineStyle.single
-                    let attributedString = NSMutableAttributedString(string: normalPrice)
-                    attributedString.addAttribute(NSAttributedString.Key.strikethroughStyle,
-                                                  value: attributedStyle.rawValue,
-                                                  range: NSRange(0..<attributedString.length))
-                    self.detailView.normalPriceLabel.attributedText = attributedString
-                }
-
-                if let discountType = self.viewModel?.discountType {
-                    self.detailView.badgeStackView.isHidden = false
-                    discountType.forEach {
-                        switch $0 {
-                        case .event:
-                            self.detailView.eventBadgeLabel.isHidden = false
-                        case .launching:
-                            self.detailView.launchingBadgeLabel.isHidden = false
-                        }
-                    }
-                }
+            DispatchQueue.main.async {[weak self] in
+                self?.navigationItem.title = self?.viewModel?.title
+                self?.detailDishSectionView.setComponents(with: self?.viewModel)
+                self?.deliverySectionView.setComponents(with: self?.viewModel)
             }
         }
     }
 
     private func setOnUpdateWithTopImages() {
         viewModel?.onUpdateWithTopImages = {
-            DispatchQueue.main.async {
-                self.viewModel?.topImages.forEach { image in
-                    let topImageView = UIImageView(image: image)
-                    topImageView.contentMode = .scaleAspectFit
-                    topImageView.widthAnchor.constraint(equalTo: topImageView.heightAnchor).isActive = true
-
-                    self.detailView.productImageContentView.addArrangedSubview(topImageView)
+            DispatchQueue.main.async {[weak self] in
+                self?.viewModel?.topImages.forEach { image in
+                    self?.detailDishSectionView.addTopImage(with: image)
                 }
             }
         }
@@ -95,20 +93,93 @@ class DetailViewController: UIViewController {
 
     private func setOnUpdateWithDetailImages() {
         viewModel?.onUpdateWithDetailImages = {
-            DispatchQueue.main.async {
-                self.viewModel?.detailImages.forEach { image in
+            DispatchQueue.main.async {[weak self] in
+                self?.viewModel?.detailImages.forEach { image in
                     let ratio = image.size.height / image.size.width
                     let detailImageView = UIImageView(image: image)
-                    let newWidth = self.detailView.detailImageStackView.frame.width
+                    let newWidth = self?.detailImageStackView.frame.width ?? 0
                     let newHeight = newWidth * ratio
 
                     detailImageView.contentMode = .scaleAspectFit
                     detailImageView.translatesAutoresizingMaskIntoConstraints = false
                     detailImageView.heightAnchor.constraint(equalToConstant: newHeight).isActive = true
 
-                    self.detailView.detailImageStackView.addArrangedSubview(detailImageView)
+                    self?.detailImageStackView.addArrangedSubview(detailImageView)
                 }
             }
         }
+    }
+}
+
+// MARK: - subView들의 layout설정
+extension DetailViewController {
+    private func setMainScrollViewLayout() {
+        let safeArea = view.safeAreaLayoutGuide
+        view.addSubview(mainScrollView)
+
+        NSLayoutConstraint.activate([
+            mainScrollView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            mainScrollView.topAnchor.constraint(equalTo: safeArea.topAnchor),
+            mainScrollView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            mainScrollView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            mainScrollView.contentLayoutGuide.topAnchor.constraint(equalTo: mainScrollView.topAnchor),
+            mainScrollView.contentLayoutGuide.leadingAnchor.constraint(equalTo: mainScrollView.leadingAnchor),
+            mainScrollView.contentLayoutGuide.widthAnchor.constraint(equalTo: mainScrollView.widthAnchor)
+        ])
+        mainScrollView.addSubview(contentView)
+
+        NSLayoutConstraint.activate([
+            contentView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            contentView.topAnchor.constraint(equalTo: mainScrollView.topAnchor),
+            contentView.bottomAnchor.constraint(equalTo: mainScrollView.bottomAnchor),
+            contentView.widthAnchor.constraint(equalTo: mainScrollView.widthAnchor)
+        ])
+    }
+
+    private func setDetailDishSectionViewLayout() {
+        detailDishSectionView.setLayout()
+        contentView.addSubview(detailDishSectionView)
+
+        NSLayoutConstraint.activate([
+            detailDishSectionView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            detailDishSectionView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            detailDishSectionView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            detailDishSectionView.widthAnchor.constraint(equalTo: contentView.widthAnchor)
+        ])
+    }
+
+    private func setDeliverySectionViewLayout() {
+        deliverySectionView.setLayout()
+        contentView.addSubview(deliverySectionView)
+
+        NSLayoutConstraint.activate([
+            deliverySectionView.topAnchor.constraint(equalTo: detailDishSectionView.bottomAnchor),
+            deliverySectionView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            deliverySectionView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor)
+        ])
+    }
+
+    private func setAmountSectionViewLayout() {
+        amountSectionView.setLayout()
+        contentView.addSubview(amountSectionView)
+
+        NSLayoutConstraint.activate([
+            amountSectionView.topAnchor.constraint(equalTo: deliverySectionView.bottomAnchor),
+            amountSectionView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            amountSectionView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            mainScrollView.contentLayoutGuide.heightAnchor.constraint(equalTo: contentView.heightAnchor)
+        ])
+    }
+
+    private func setDetailImageStackViewLayout() {
+        contentView.addSubview(detailImageStackView)
+
+        NSLayoutConstraint.activate([
+            detailImageStackView.topAnchor.constraint(equalTo: amountSectionView.bottomAnchor, constant: 48),
+            detailImageStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            detailImageStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            detailImageStackView.bottomAnchor.constraint(equalTo: mainScrollView.bottomAnchor),
+            contentView.bottomAnchor.constraint(equalTo: detailImageStackView.bottomAnchor)
+        ])
     }
 }
